Small refactor and clean-up of headers
