+++ libtiledbsoma/include/tiledbsoma/arrow_adapter.h
@@ -2,7 +2,7 @@
 #define ARROW_ADAPTER_H
 
 #include <tiledbsoma/tiledbsoma>
-#ifndef NO_CARROW_INCLUDE
+#ifndef ARROW_SCHEMA_AND_ARRAY_DEFINED 
 #include "carrow.h"
 #endif
